remove base JsonApiFormatter
