+++ src/JsonApiDotNetCore/Middleware/JsonApiOutputFormatter.cs
@@ -1,13 +1,21 @@
+using System;
+using System.Collections;
+using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
+using JsonApiDotNetCore.Resources;
 using JsonApiDotNetCore.Serialization;
+using Microsoft.AspNetCore.Mvc.ApiExplorer;
 using Microsoft.AspNetCore.Mvc.Formatters;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Net.Http.Headers;
 
 namespace JsonApiDotNetCore.Middleware
 {
-    /// <inheritdoc cref="JsonApiDotNetCore.Middleware.IJsonApiInputFormatter" />
-    public sealed class JsonApiOutputFormatter : JsonApiFormatter, IJsonApiOutputFormatter
+    public sealed class JsonApiOutputFormatter : IJsonApiOutputFormatter, IApiRequestFormatMetadataProvider
     {
+        private static readonly Type OperationContainerType = typeof(OperationContainer);
+
         /// <inheritdoc />
         public bool CanWriteResult(OutputFormatterCanWriteContext context)
         {
@@ -23,6 +31,25 @@ namespace JsonApiDotNetCore.Middleware
 
             var writer = context.HttpContext.RequestServices.GetRequiredService<IJsonApiWriter>();
             await writer.WriteAsync(context);
+        }
+
+        /// <inheritdoc />
+        public IReadOnlyList<string> GetSupportedContentTypes(string contentType, Type objectType)
+        {
+            ArgumentGuard.NotNull(contentType, nameof(contentType));
+            ArgumentGuard.NotNull(objectType, nameof(objectType));
+
+            string mediaType = IsAtomicOperationsType(objectType) ? HeaderConstants.AtomicOperationsMediaType : HeaderConstants.MediaType;
+
+            return new MediaTypeCollection
+            {
+                new MediaTypeHeaderValue(mediaType)
+            };
+        }
+
+        private bool IsAtomicOperationsType(Type objectType)
+        {
+            return objectType.GetInterface(nameof(IEnumerable)) != null && objectType.GetGenericArguments().First() == OperationContainerType;
         }
     }
 }
